Changes needed for randmath to run somehow. The body of the benchmark was previously not being called. Also, without one of the variables being "volatile", it almost all got optimised away. Also, tohost was to narrow, so a write to the neighbouring global variable ended the test; I've made it long long now.

+++ syscalls.c
@@ -19,7 +19,7 @@ _start_purecap(void) {
 }
 #endif
 
-extern volatile int tohost = 0;
+extern volatile long long tohost = 0;
 
 int main(int, char **);
